+++ src/rockverse/_assert/condition.py
@@ -13,9 +13,9 @@ def non_negative_integer_or_float(varname, var):
     collective_raise(ValueError(f"Expected non negative number for {varname}."))
 
 def integer(varname, var):
-    if np.dtype(type(var)).kind in 'ui': #splitting to handle >= op only if number
+    if np.dtype(type(var)).kind in 'ui':
         return
-    collective_raise(ValueError(f"Expected non negative integer for {varname}."))
+    collective_raise(ValueError(f"Expected integer for {varname}."))
 
 def non_negative_integer(varname, var):
     if np.dtype(type(var)).kind in 'ui': #splitting to handle >= op only if number